Split add/edit organization to own e2e test file
